Update CoC contact email
We previously used unicorn-utterances simply due to an existing wildcard email forward being setup and easy to use

I thought (wrongly) that it was going to be equally as difficult to setup an email for our new domain. Turns out it was like two clicks 🙃

To be crystal clear: UU is not directly in part of (or in charge of) The Frame Drops. While some of the admins are part of both communities,
they will be treated as distinct entitites and absolutely no promotion of UU will occur from The Frame Drops (outside of charity-related events &
individual streamer commands [like !discord] that may link to it)