Implement temporary workaround to inaccessible calendar comp
This is a super unfortunate result of using Vuetify. We have super limited time in order to get the site up-and-running before the event.

This was the best short-measurement we could do to make the events somewhat accessible given our current constraints. I'm so so sorry.

We plan on doing more of these events in the future, and I pledge to have this properly sorted by the next event we do